% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset_detection.R
\name{onset_detection}
\alias{onset_detection}
\title{Onset detection}
\usage{
onset_detection(
  file_name,
  transform_file = NULL,
  normalise = 1L,
  hidePrint = TRUE,
  type = NULL,
  if_bad_result_return_single_na = NULL,
  tidy_fun = tidy_onset_detection
)
}
\arguments{
\item{tidy_fun}{}
}
\description{
Onset detection
}
